Update printers.py
Depending on the topology of the red-black tree, the right-hand subtrees are not parsed correctly and repeated entries are printed in gdb.
These proposed changes match with the c++ code implementation of libc++::map::iterator  
example of a map that produces such tree is posted below. set a breakpoint in gdb on the second for loop and print data. you'll see the wrong output

#include <map>
#include <string>
#include <iostream>

std::map<std::string, int> data;

const char *names[] = {
  "toop",
  "boogh",
  "khyar",
  "goje",
  "jooje",
  "saat",
  "cheragh",
  "ghalam",
  "medad",
  "kagaz"
};

int main (void) {
  for (int i = 0; i<10; i++) {
    data[std::string(names[i])] = i;
  }

  for (auto m : data) {
    std::cout << m.first <<" "<<m.second <<"\n";
  }
  
  return 0;
}